requirements.txt
Flask==2.3.4
Flask-Mail==0.9.1
Flask-WTF==1.1.1
python-dotenv==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Werkzeug==2.3.7